+++ ethcore/src/json_tests/chain.rs
@@ -20,6 +20,7 @@ use pod_state::*;
 use block::Block;
 use ethereum;
 use tests::helpers::*;
+use devtools::*;
 
 pub fn json_chain_test(json_data: &[u8], era: ChainEra) -> Vec<String> {
 	init_log();
