+++ ethcore/src/lib.rs
@@ -92,6 +92,7 @@ extern crate env_logger;
 extern crate num_cpus;
 extern crate crossbeam;
 
+#[cfg(test)] extern crate ethcore_devtools as devtools;
 #[cfg(feature = "jit" )] extern crate evmjit;
 
 pub mod block;
